Servant.OAuth2

Description

A modern wrapper around https://github.com/fpco/wai-middleware-auth/.

This library can be used to add type-level authentication and authorisation into a Servant api via OAuth2. You might be interested in this if you are using Servant as a light-weight webserver serving Html, for example.

Documentation

data Tag a (rs :: [Type])

A simple way to add a type-level tag onto the return type for your AuthServerData instance. Used something like: Tag Github ...; this then allows you to use multiple oauth providers on one server, and have servant still pick out the right auth handler to use. See: https://docs.servant.dev/en/stable/tutorial/Authentication.html#recap for more detail.

Since: 0.1.0.0

Constructors

Tag

Fields

unTag :: Union rs

type Ident = ByteString

This is the result of successful completion of the OAuth2 login workflow; it is the identifier that comes back from the provider.

Since: 0.1.0.0

data OAuth2Routes (rs :: [Type]) mode

This contains the complete route that the given OAuth2 provider will return to. This implementation for this is fully given by the authServer function.

Since: 0.1.0.0

Constructors

AuthRoutes

Fields

complete :: mode :- ("complete" :> UVerb 'GET '[HTML] rs)

authServer :: forall m a (rs :: [Type]). Monad m => Tag a rs -> OAuth2Routes rs (AsServerT m)

The server implementation for the OAuth2Routes routes. Ultimately, this just returns the result of the success function from OAuth2Settings.

Since: 0.1.0.0

oauth2AuthHandler :: forall m p rs e. (AuthProvider p, MonadIO m, MonadThrow m, MonadError e m, Monad m) => OAuth2Settings m p rs -> (m (Tag p rs) -> Handler (Tag p rs)) -> AuthHandler Request (Tag p rs)

The central handler that runs when the complete route is called. In here we pass of to handleLogin via runOAuth2, and we unwrap the results; if there was any error, we throw a servant error, or, in the happy case that we successfully authenticate, we call the success function and return, which is then (after unwrapping) returned by the authServer.

Since: 0.1.0.0

runOAuth2 :: (MonadIO m, AuthProvider p) => Request -> p -> (AuthLoginState -> IO Response) -> (Status -> ByteString -> IO Response) -> m Response

In the context of Wai, run the complete step of the OAuth2 process. We return a Response, unfortunately, which we will later interpret into Servant responses.

Since: 0.1.0.0

data OAuth2Settings m p (rs :: [Type])

Used to record the particular provider you are using, along with the ultimate return type of the complete route, that will, in the end, need to agree with the particular implementation of the success function.

Since: 0.1.0.0

Constructors

OAuth2Settings

Fields

success :: Request -> Ident -> m (Union rs)

provider :: p

defaultOAuth2Settings :: forall m p. Applicative m => p -> OAuth2Settings m p '[WithStatus 200 Text]

Default settings, only really suitable for demo purposes, that simply always respond with just the return value of the OAuth2 route; typically the email, depending on how your provider is configured.

Note that in order to use this, your instance of AuthServerData must return '[WithStatus 200 Text].

Since: 0.1.0.0