Servant.OAuth2.Cookies

Documentation

type RedirectWithCookie = Headers '[Header "Location" Text, Header "Set-Cookie" SetCookie] NoContent

A helpful alias for returning a Location header and a SetCookie header.

Since: 0.1.0.0

ourCookie :: ByteString

Default name of our cookie.

Since: 0.1.0.0

emptyCookie :: SetCookie

An empty cookie; useful to set when logging out.

Since: 0.1.0.0

redirectWithCookie :: Text -> SetCookie -> RedirectWithCookie

Set a cookie and then perform a redirection; typically used as part of logging in; i.e. after successfully performing OAuth2 authentication, we just want to redirect back to the homepage.

Since: 0.1.0.0

simpleCookieOAuth2Settings :: forall s m p. (Binary s, Applicative m, Monad m, MonadIO m) => p -> (Request -> Ident -> m s) -> Key -> OAuth2Settings m p '[WithStatus 303 RedirectWithCookie]

Build a simple cookie provided you have a function that can convert the ident into a sessionId kind of object.

Since: 0.1.0.0

buildSessionCookie :: Binary s => Key -> s -> IO SetCookie

Make a session cookie from the ident; i.e. just set the cookie to be the ident value.

Since: 0.1.0.0

getSessionIdFromCookie :: Binary s => Request -> Key -> Maybe s

Perform the decryption of the cookie in reverse order to buildSessionCookie.

Since: 0.1.0.0