Servant.OAuth2.Hacks

Description

A collection of hacky functions needed to work around the fact that 'wai-middleware-auth' is presently defined in terms of Wai, and, really, we want something that is defined for Servant.

Documentation

getGithubLoginUrl :: Text -> OAuth2Settings m Github a -> Text

For some settings specialised to Github, return the login url.

Since: 0.1.0.0

getGoogleLoginUrl :: Text -> OAuth2Settings m Google a -> Text

For some settings specialised to Google, return the login url.

Since: 0.1.0.0

getRedirectUrl :: Text -> OAuth2 -> Maybe [Text] -> Text

An extremely unfortunate way of getting the redirect URL; stolen from Internal.

Since: 0.1.0.0